Linted

+++ game_core/src/game_engine/engine.rs
@@ -1,4 +1,4 @@
-use crate::{constants::{INITIAL_CAMERA_VIEWPORT, TILE_SIZE, WORLD_ID_NONE}, dialogues::{menu::DialogueMenu, models::Dialogue}, features::{creep_spawner::CreepSpawner, death_screen::DeathScreen, destination::Destination, loading_screen::LoadingScreen}, menus::{confirmation::ConfirmationDialog, entity_options::EntityOptionsMenu, game_menu::GameMenu, long_text_display::LongTextDisplay, toasts::{Toast, ToastDisplay}}, ui::components::Typography, utils::{rect::Rect, vector::Vector2d}};
+use crate::{constants::{INITIAL_CAMERA_VIEWPORT, TILE_SIZE, WORLD_ID_NONE}, dialogues::{menu::DialogueMenu, models::Dialogue}, features::{creep_spawner::CreepSpawner, death_screen::DeathScreen, destination::Destination, loading_screen::LoadingScreen}, menus::{confirmation::ConfirmationDialog, entity_options::EntityOptionsMenu, game_menu::GameMenu, long_text_display::LongTextDisplay, toasts::{Toast, ToastDisplay}}, utils::{rect::Rect, vector::Vector2d}};
 
 use super::{inventory::{add_to_inventory, remove_from_inventory}, keyboard_events_provider::{KeyboardEventsProvider, NO_KEYBOARD_EVENTS}, mouse_events_provider::MouseEventsProvider, state_updates::{EngineStateUpdate, WorldStateUpdate}, storage::{get_value_for_key, set_value_for_key, StorageKey}, world::World};
 
@@ -145,8 +145,8 @@ impl GameEngine {
     ) {
         self.camera_viewport.w = (width / (scale * TILE_SIZE)) as i32;
         self.camera_viewport.h = (height / (scale * TILE_SIZE)) as i32;
-        self.long_text_display.max_line_length = (width as f32 / font_size).floor() as usize;
-        self.long_text_display.visible_line_count = (0.3 * height as f32 / (line_spacing + font_size)).floor() as usize;
+        self.long_text_display.max_line_length = (width / font_size).floor() as usize;
+        self.long_text_display.visible_line_count = (0.3 * height / (line_spacing + font_size)).floor() as usize;
     }
 
     pub fn rendering_scale_for_screen_width(&self, width: f32) -> (f32, f32) {
